LayoutTests/canvas/philip/tests/2d.transformation.scale.multiple-expected.txt

< [index] >
2d.transformation.scale.multiple
Multiple scale()s combine
References: 2d.transformation.scale.multiple
Actual output:
Expected output:

Passed
